Перетусовал папки и файлы для удобства
