--- app/src/main/java/ru/practicum/android/diploma/data/dto/VacancyDTO.kt
@@ -1,9 +1,0 @@
-package ru.practicum.android.diploma.data.dto
-
-data class VacancyDTO(
-    val id: String,
-    val name: String,
-    val employer: String,
-    val area: String,
-    val salary: String
-)
