--- app/src/main/java/ru/practicum/android/diploma/data/network/Response.kt
@@ -1,5 +1,0 @@
-package ru.practicum.android.diploma.data.network
-
-open class Response {
-    var resultCode = 0
-}
