--- app/src/main/java/ru/practicum/android/diploma/data/network/VacanciesResponse.kt
@@ -1,5 +1,0 @@
-package ru.practicum.android.diploma.data.network
-
-import ru.practicum.android.diploma.data.dto.VacancyDTO
-
-data class VacanciesResponse(val results: ArrayList<VacancyDTO>) : Response()
