--- app/src/main/java/ru/practicum/android/diploma/data/network/VacancySearchRequest.kt
@@ -1,3 +1,0 @@
-package ru.practicum.android.diploma.data.network
-
-data class VacancySearchRequest(val request: String)
