+++ app/src/main/java/ru/practicum/android/diploma/data/search/dto/VacancyDTO.kt
@@ -1,0 +1,9 @@
+package ru.practicum.android.diploma.data.search.dto
+
+data class VacancyDTO(
+    val id: String,
+    val name: String,
+//    val employer: String,
+//    val area: String,
+//    val salary: String
+)
