+++ app/src/main/java/ru/practicum/android/diploma/data/search/network/NetworkClient.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.data.network
+package ru.practicum.android.diploma.data.search.network
 
 interface NetworkClient {
     suspend fun doRequest(dto: Any): Response
