+++ app/src/main/java/ru/practicum/android/diploma/data/search/network/VacanciesResponse.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.data.search.network
+
+import ru.practicum.android.diploma.data.search.dto.VacancyDTO
+
+data class VacanciesResponse(val results: ArrayList<VacancyDTO>) : Response()
