+++ app/src/main/java/ru/practicum/android/diploma/data/search/network/VacancySearchRequest.kt
@@ -1,0 +1,3 @@
+package ru.practicum.android.diploma.data.search.network
+
+data class VacancySearchRequest(val request: String)
