+++ app/src/main/java/ru/practicum/android/diploma/di/search/SearchDataModule.kt
@@ -6,9 +6,9 @@ import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import ru.practicum.android.diploma.data.network.AppAPI
-import ru.practicum.android.diploma.data.network.NetworkClient
-import ru.practicum.android.diploma.data.network.RetrofitNetworkClient
+import ru.practicum.android.diploma.data.search.network.AppAPI
+import ru.practicum.android.diploma.data.search.network.NetworkClient
+import ru.practicum.android.diploma.data.RetrofitNetworkClient
 import ru.practicum.android.diploma.data.search.SearchRepositoryImpl
 import ru.practicum.android.diploma.data.utils.InternetAccessChecker
 import ru.practicum.android.diploma.domain.search.SearchInteractor
@@ -19,7 +19,7 @@ val searchDataModule = module {
     single { InternetAccessChecker() }
     single<AppAPI> {
         Retrofit.Builder()
-            .baseUrl("https://api.hh.ru/")
+            .baseUrl("https://api.hh.ru")
             .addConverterFactory(GsonConverterFactory.create())
             .build()
             .create(AppAPI::class.java)
@@ -29,14 +29,6 @@ val searchDataModule = module {
 
     factory<NetworkClient> {
         RetrofitNetworkClient(get(), get())
-    }
-
-    single<SearchRepository> {
-        SearchRepositoryImpl(get())
-    }
-
-    single<SearchInteractor> {
-        SearchInteractorImpl(get())
     }
 
     single { (key: String) ->
