+++ app/src/main/java/ru/practicum/android/diploma/util/Constants.kt
@@ -1,11 +1,8 @@
 package ru.practicum.android.diploma.util
 
-import ru.practicum.android.diploma.BuildConfig
-
 const val EMPTY_STRING = ""
 
 // ---- Network const ----
-const val ACCESS_TOKEN = BuildConfig.HH_ACCESS_TOKEN
 const val HH_REQUEST_HEADER = "Almighty job seeker/1.0 (andrei.urich@yandex.ru)"
 const val RESULT_CODE_BAD_REQUEST = 400
 const val RESULT_CODE_ERROR = -1
